fix: Deprecate OC_Template, add proper template manager instead
